Document debugging in RustRover.
Adds in documentation for Debugging within RustRover, without having to
use `--no-default-features` and thus with dynamic linking.

While PR #362 adds in a Run Configuration allowing for an easy and quick
debugging experience, it also has to disable features, to disable
dynamic linking. As a side effect, this disables several useful tools.

Unfortunately, we can't implement this directly in the template, because
there are paths that are dependent upon the person using this. As such
we provide this documentation so users can set it up on their own.